// src/app/page/master/pekerjaan/pekerjaan.interface.ts
import { EventEmitter } from '@angular/core';

export interface Pekerjaan {
    kode?: any;
    namaPemilikProfile?: string;
    reportDisplay?: string;
    kodeExternal?: string;
    namaExternal?: string;
    kdKelompokPemilik?: boolean;
    statusEnabled?: boolean;
    noRec?: string;
    label?: any;
    value?: any;
    kelompokPemilik?: string;
    id?: any;
}